修复直播清晰度切换问题

+++ src/Models/Models.BiliBili/User/ChatSessionListResponse.cs
@@ -167,7 +167,7 @@ public class BiliChatMsg
     /// 接收方 Id.
     /// </summary>
     [JsonPropertyName("receiver_id")]
-    public int ReceiverId { get; set; }
+    public long ReceiverId { get; set; }
 
     /// <summary>
     /// 内容.
